zkrune-todo-bot/requirements.txt
python-telegram-bot==21.10
supabase==2.13.0
